Moves window.px.card and window.px.dealer behaviors to a single, cleaned up file
This commit extracts the window.px.card behavior from px-card.html and the
window.px.dealer behavior from the deprecated px repo and places them in
a single cleaned up behavior file. This package might be useful for teams
that want to continue using these behaviors when they are deprecated and
removed from px-card in the next major revision of the component.